apirequirements.txt
robotframework-appiumlibrary
robotframework-databaselibrary
robotframework-debuglibrary
robotframework-jsonlibrary
robotframework-faker
allure-robotframework
robotframework-requests
robotframework-selenium2library
robotframework-httplibrary
robotframework-allurereport
scikit-build
robotframework-screencaplibrary
robotframework-jsonvalidator
gherkin2robotframework
robotframework-xvfb
robotframework-lint
requests
robotframework
jsondiff
